+++ src/engine/Graphics/Context/debug-line-renderer/debug-line-vertex.glsl
@@ -1,8 +1,10 @@
 #version 300 es
 in vec2 a_position;
 in vec4 a_color;
+in float a_lengthSoFar;
 
 out lowp vec4 v_color;
+out float v_lengthSoFar;
 
 uniform mat4 u_matrix;
 
@@ -12,4 +14,7 @@ void main() {
 
    // Passthrough the color
    v_color = a_color;
+
+   // Passthrough the line length so far
+   v_lengthSoFar = a_lengthSoFar;
 }